# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TravelPlace {
  id: string;
  name: string;
  category: string;
  /** @nullable */
  address: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  website: string | null;
  mapLink: string;
  source: string;
  /** @nullable */
  price: string | null;
  availability: string;
}

export interface TravelPlacesResponse {
  city: string;
  source: string;
  retrievedAt: string;
  places: TravelPlace[];
}

export interface TravelAirport {
  id: string;
  city: string;
  country: string;
  airportName: string;
  iataCode: string;
  latitude: number;
  longitude: number;
  source: string;
}

export interface TravelAirportsResponse {
  query: string;
  source: string;
  airports: TravelAirport[];
}

export interface TravelWeatherDay {
  date: string;
  weatherCode: number;
  minC: number;
  maxC: number;
  precipitationProbability: number;
}

export interface TravelWeatherResponse {
  city: string;
  source: string;
  retrievedAt: string;
  latitude: number;
  longitude: number;
  days: TravelWeatherDay[];
  /** @nullable */
  note?: string | null;
}

export interface TravelSearchLinksResponse {
  origin: string;
  destination: string;
  flightSearchUrl: string;
  accommodationSearchUrl: string;
  source: string;
  note: string;
}

export interface TravelExchangeRateResponse {
  from: string;
  to: string;
  rate: number;
  source: string;
  retrievedAt: string;
}

export type SearchTravelPlacesParams = {
/**
 * @minLength 2
 */
city: string;
kind?: string;
/**
 * @minimum 1
 * @maximum 30
 */
limit?: number;
};

export type SearchTravelAirportsParams = {
/**
 * @minLength 2
 */
q: string;
/**
 * @minimum 1
 * @maximum 10
 */
limit?: number;
};

export type GetTravelWeatherParams = {
/**
 * @minLength 2
 */
city: string;
startDate?: string;
endDate?: string;
};

export type GetTravelSearchLinksParams = {
/**
 * @minLength 2
 */
origin: string;
/**
 * @minLength 2
 */
destination: string;
departureDate?: string;
returnDate?: string;
/**
 * @minimum 1
 * @maximum 12
 */
travelers?: number;
};

export type GetTravelExchangeRateParams = {
/**
 * @minLength 3
 * @maxLength 3
 */
from: string;
/**
 * @minLength 3
 * @maxLength 3
 */
to: string;
};