# lib/api-zod/src/generated/types/travelAirport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelAirport {
  id: string;
  city: string;
  country: string;
  airportName: string;
  iataCode: string;
  latitude: number;
  longitude: number;
  source: string;
}