# lib/api-zod/src/generated/types/travelAirportsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TravelAirport } from './travelAirport';

export interface TravelAirportsResponse {
  query: string;
  source: string;
  airports: TravelAirport[];
}