# lib/api-zod/src/generated/types/travelExchangeRateResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelExchangeRateResponse {
  from: string;
  to: string;
  rate: number;
  source: string;
  retrievedAt: Date;
}